requirements.txt
asgiref==3.5.2
black==22.10.0
chardet==5.0.0
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-cors-headers==3.13.0
django-csvimport==2.16
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
tzlocal==4.2
